+++ src/lib.rs
@@ -12,9 +12,8 @@ pub enum Error {
     Dfu(#[from] dfu_core::Error),
     #[error(transparent)]
     Io(#[from] std::io::Error),
-    // TODO: it's kinda annoying to have a lifetime in the error
     #[error("Could not parse memory layout: {0}")]
-    MemoryLayout(String),
+    MemoryLayout(dfu_core::memory_layout::Error),
     #[error("rusb: {0}")]
     LibUsb(#[from] rusb::Error),
     #[error("The device has no languages.")]
@@ -141,7 +140,7 @@ impl<C: rusb::UsbContext> DfuLibusb<C> {
                 .rsplit_once('/')
                 .ok_or(Error::InvalidInterfaceString)?;
             let memory_layout = dfu_core::memory_layout::MemoryLayout::try_from(memory_layout)
-                .map_err(|err| Error::MemoryLayout(err.to_string()))?;
+                .map_err(Error::MemoryLayout)?;
             let (_rest, address) = rest.rsplit_once('/').ok_or(Error::InvalidInterfaceString)?;
             let address = address
                 .strip_prefix("0x")