+++ CHANGELOG.md
@@ -4,3 +4,5 @@ Changelog
 ## v0.2.0
 
 - Moved cli binary to example `download.rs`
+- Change error type for memory layout from String to core's memory layout error
+  type